+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to **Device Detector** are documented in this *changelog*.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **Device Detector** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased - will be 1.1.2]
+### Removed
+- Compatibility with WordPress versions prior to 5.2.
 
 ## [1.1.1] - 2020-02-16
 ### Changed